completed edit, working on task complete functionality
