+++ src/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" href="./styles/main.css">
-    <title>Document</title>
+    <title>Sons of Witches (Nutshell)</title>
 </head>
 
 <body>
@@ -16,54 +16,18 @@
             <section class="auth--register"></section>
         </article>
         <article class="dashboard">
-            <!-- All your Nutshell are belong here -->
-            <section class="user">
 
+            <section class="user"></section>
 
+            <section class="events"></section>
 
+            <section class="tasks"></section>
 
+            <section class="chat"></section>
 
-            </section>
+            <section class="friends"></section>
 
-            <section class="events">
-
-
-
-
-
-            </section>
-
-            <section class="tasks">
-
-
-
-
-
-            </section>
-
-            <section class="chat">
-
-
-
-
-
-            </section>
-
-            <section class="friends">
-
-
-
-
-
-            </section>
-
-            <section class="articles">
-
-
-
-
-
-            </section>
+            <section class="articles"></section>
 
         </article>
     </article>
