+++ src/scripts/Tasks/TasksForm.js
@@ -1,28 +1,23 @@
-import { saveTasks } from "./TasksDataProvider.js"
+import { newContentTarget } from "./TasksList.js"
 
-const eventHub = document.querySelector(".container")
 const contentTarget = document.querySelector(".tasks")
 
-
-
-/* Function that adds a form to the display.
-*/
 export const renderTasksForm = () => {
     contentTarget.innerHTML = `
     <h2>Create a Task:</h2>
     <div class="tasksForm">
         <fieldset>
-            <label>Title</label>
+            <label>Title:</label>
             <input type="text" id="tasksTitle">
         </fieldset>
 
         <fieldset>
-            <label>Summary</label>
+            <label>Summary:</label>
             <input type="text" id="tasksSummary">
         </fieldset>
 
         <fieldset>
-            <label>Due Date</label>
+            <label>Due Date:</label>
             <input type="text" id="tasksDueDate">
         </fieldset>
 
@@ -41,20 +36,20 @@ export const renderTasksForm = () => {
 `
 }
 
-
-/* Event Listener that will pass an object to the saveTasks
-   function if the target.id is equal to the ID of tasksSaveBtn
-*/ 
-   
-eventHub.addEventListener("click", event => {
-    if(event.target.id === "tasksSaveBtn"){
-        const newTasksObj = {
-            userId: parseInt(sessionStorage.getItem("userId")),
-            title: document.querySelector("#tasksTitle").value,
-            summary: document.querySelector("#tasksSummary").value,
-            date: document.querySelector("#tasksDueDate").value,
-            complete: false
-        }
-        saveTasks(newTasksObj)
-    }
-})
+export const editForm = (tasksObj) => {
+    newContentTarget.innerHTML = `
+         <div class="formPopup" id="tasksEditForm">
+            <div class="editFormContainer">
+                <h3>Edit Task</h3>
+                    <lable>Title</label>
+                    <input type="text" id="editTasksTitle" value="${tasksObj.title}"></input>
+                    <lable>Summary</label>
+                    <input type="text" id="editTasksSummary" value="${tasksObj.summary}"></input>
+                    <lable>Due Date</label>
+                    <input type="text" id="editTasksDate" value="${tasksObj.date}"></input>
+                    <button type="submit" id="editBtn--${tasksObj.id}">Save</button>
+                    <button type="button" id="editCancel">Cancel</button>
+            </div>
+        </div> 
+    `
+}
