+++ src/scripts/Tasks/TasksList.js
@@ -1,23 +1,15 @@
-import { getTasks, useTasks, deleteTasks, editTasks } from "./TasksDataProvider.js"
+import { saveTasks, getTasks, useTasks, deleteTasks, editTasks, completeTasks } from "./TasksDataProvider.js"
+import { editForm } from "./tasksForm.js"
 
 const eventHub = document.querySelector(".container")
-const contentTarget = document.querySelector(".tasks")
-let contentTarget1
 
-/* Function that maps through an array then returns that array as the 
-    argument for the function tasksHTML
-*/    
-const render = () => {
-    const contentTarget = document.querySelector("#tasksList")
-    contentTarget.innerHTML = useTasks().map(tasks => {
-        return tasksHTML(tasks)
-    }).join("")
+export let newContentTarget
+ 
+export const tasksList = () => {
+    getTasks()
+    .then(render)
 }
 
-/* Function that takes the argument from the rendor function
-    and creates the HTML useing the information that was passed
-    as the argument.
-*/    
 const tasksHTML = (tasksObj) => {
     return `
         <div>Title: ${tasksObj.title}</div>
@@ -28,18 +20,21 @@ const tasksHTML = (tasksObj) => {
     `
 }
 
-/* Function that adds a delete button to each task in the list
-   but only if the session user is equal to the user that posted
-   the task.
-*/      
+const render = () => {
+    const contentTarget = document.querySelector("#tasksList")
+    contentTarget.innerHTML = useTasks().map(tasks => {
+        return tasksHTML(tasks)
+    }).join("")
+}
+      
 const checkUserId = (tasks) => {
     let userId = sessionStorage.getItem("userId")
     if(parseInt(userId) === tasks.userId){
         return `
-        <button id="deleteTasks--${tasks.id}">Delete</button>
-        <button id="editTasks--${tasks.id}">Edit</button>
-        <button id="completeTasks--${tasks.id}">Task Complete</button>
-        <div id="editFormTarget--${tasks.id}"></div>
+            <button id="deleteTasks--${tasks.id}">Delete</button>
+            <button id="editTasks--${tasks.id}">Edit</button>
+            <button id="completeTasks--${tasks.id}">Task Complete</button>
+            <div id="editFormTarget--${tasks.id}"></div>
         `
     }
     else{
@@ -47,18 +42,19 @@ const checkUserId = (tasks) => {
     }
 }
 
-/* Function exported to main.js that is responsible for calling the list to 
-   be added to the DOM
-*/   
-export const tasksList = () => {
-    getTasks()
-    .then(render)
-}
+eventHub.addEventListener("click", event => {
+    if(event.target.id === "tasksSaveBtn"){
+        const newTasksObj = {
+            userId: parseInt(sessionStorage.getItem("userId")),
+            title: document.querySelector("#tasksTitle").value,
+            summary: document.querySelector("#tasksSummary").value,
+            date: document.querySelector("#tasksDueDate").value,
+            complete: false
+        }
+        saveTasks(newTasksObj)
+    }
+})
 
-/* Event listener that will check if the delete button was clicked.
-   If the delete button was clicked then the id to be deleted will
-   be passed to deleteTasks.
- */  
 eventHub.addEventListener("click", event => {
     if(event.target.id.startsWith("deleteTasks")){
         const [prefix, id] = event.target.id.split("--")
@@ -66,47 +62,17 @@ eventHub.addEventListener("click", event => {
     }
 })
 
-/* Event Listener that checks if the edit button is clicked.
-    If edit was clicked it calls the tasksEditForm function and passes
-    the object to be edited.
-*/
 eventHub.addEventListener("click", event => {
     if (event.target.id.startsWith("editTasks")) {
-        console.log("1")
         const [prefix, id] = event.target.id.split("--")
         let x = useTasks().find(tasks => {
             return parseInt(tasks.id) === parseInt(id)
         })
-        contentTarget1 = document.querySelector(`#editFormTarget--${id}`)
+        newContentTarget = document.querySelector(`#editFormTarget--${id}`)
         editForm(x)
     }
 })
 
-/* Function that takes an object and displays the object it editable fields
-    in the HTML at a target location
- */
-const editForm = (tasksObj) => {
-    contentTarget1.innerHTML = `
-         <div class="formPopup" id="tasksEditForm">
-            <div class="editFormContainer">
-                <h3>Edit Task</h3>
-                    <lable>Title</label>
-                    <input type="text" id="editTasksTitle" value="${tasksObj.title}"></input>
-                    <lable>Summary</label>
-                    <input type="text" id="editTasksSummary" value="${tasksObj.summary}"></input>
-                    <lable>Due Date</label>
-                    <input type="text" id="editTasksDate" value="${tasksObj.date}"></input>
-                    <button type="submit" id="editBtn--${tasksObj.id}">Save</button>
-                    <button type="button" id="editCancel">Cancel</button>
-            </div>
-        </div> 
-    `
-}
-
-/* Event listener that checks if save is clicked in the edit form is cliked.
-    If the save button is clicked on the edit form it will call the editArticle
-    function and pass the information.
- */
 eventHub.addEventListener("click", event => {
     if(event.target.id.startsWith("editBtn")){
         const [prefix, id] = event.target.id.split("--")
@@ -120,9 +86,6 @@ eventHub.addEventListener("click", event => {
     }
 })
 
-/* Event listener that checks if the cancel edit button is click.
-    If the button is clicked it will call the tasksList function.
- */
 eventHub.addEventListener("click", event => {
     if(event.target.id === "editCancel"){
         tasksList()
@@ -131,18 +94,8 @@ eventHub.addEventListener("click", event => {
 
 
 
+eventHub.addEventListener("click", event => {
+    if(event.target.id === "taskComplete"){
 
-
-
-
-
-/* Event listener that will check if the task complete button was clicked.
-   If the task complete button was clicked then the id to be edited will
-   be passed to completeTasks.
- */  
-// eventHub.addEventListener("click", event => {
-//     if(event.target.id.startsWith("completeTasks")){
-//         const [prefix, id] = event.target.id.split("--")
-//         editTasks(id)
-//     }
-// })
+    }
+})
